making search page its own route. This allows me to take off a lot of redux state dependencies
