+++ frontend/components/SearchPanelComponents/StandardSearch/StandardSearch.js
@@ -1,5 +1,4 @@
 import axios from 'axios';
-import { useRouter } from 'next/router';
 import { useEffect, useRef, useState } from 'react';
 import Loader from 'react-loader-spinner';
 import { useDispatch, useSelector } from 'react-redux';
@@ -27,7 +26,6 @@ export default function StandardSearch() {
   const [firstQuery, setFirstQuery] = useState(true);
   const hasQueryChanged = useCompare(query);
   const [count, setCount] = useState();
-  const router = useRouter();
 
   const { newCount, isCountLoading, isCountError } = useSearchCount(
     query,
@@ -36,9 +34,6 @@ export default function StandardSearch() {
   );
 
   useEffect(() => {
-    router.push(`/?search=${query}&offset=${offset}`, undefined, {
-      shallow: true
-    });
     if (hasQueryChanged && !firstQuery) {
       dispatch(setOffset(0));
     }
