+++ frontend/pages/search.js
@@ -1,0 +1,5 @@
+import StandardSearch from './search/[...query]';
+
+export default function SearchDefault() {
+  return <StandardSearch />;
+}
