+++ frontend/pages/search/[...query].js
@@ -1,0 +1,10 @@
+import SearchPanel from '../../components/SearchPanel';
+import TopLevel from '../../components/TopLevel';
+
+export default function StandardSearch() {
+  return (
+    <TopLevel searchMode={true}>
+      <SearchPanel />
+    </TopLevel>
+  );
+}
